Extracted shared sysbone aiming math into EveTurretAiming
